refactor: expose lifespan checks
